backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pyyaml>=6.0.1
chardet>=5.2.0
langchain-text-splitters>=0.0.1
